+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## next
 
-- Add a new options: `initialViewIndex` (#183).
+- Add 2 new options: `toggleOnDblclick` (#173) and `initialViewIndex` (#183).
 - Enhance the `title` option to support to customize title content (#54, #185).
 
 ## 1.1.0 (May 27, 2018)